*(add|modify:添加move2_subctrl使能|调整J1，J3的角度映射|实现基本自定义控制器控制)

+++ components/devices/Custom_contrllor/Custom_ctrl.h
@@ -42,7 +42,7 @@ typedef __packed struct
 	frame_header_t	header;
 	uint16_t cmd_id;
 	key_t   key;
-	uint16_t adc_val[5];
+	float adc_val[5];
 	uint16_t crc16;
 } data_t;
